@@ -3,7 +3,8 @@ package com.beads.model.builder;
 import com.beads.model.domain.Order;
 import com.beads.model.domain.OrderItem;
 import com.beads.model.domain.OrderStatus;
-import org.joda.time.DateTime;
+
+import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
 
@@ -30,7 +31,7 @@ public class OrderBuilder {
         withOrderDetails("orderDetails");
         withStatus(OrderStatus.PENDING);
         withPhoneNumber("1234567890");
-        withModifiedDate(DateTime.now());
+        withModifiedDate(Instant.now());
         withOrderItems(buildOrderItems());
     }
 
@@ -39,7 +40,7 @@ public class OrderBuilder {
         return this;
     }
 
-    private OrderBuilder withModifiedDate(DateTime modifiedDate) {
+    private OrderBuilder withModifiedDate(Instant modifiedDate) {
         order.setModifyDate(modifiedDate);
         return this;
     }
